+++ utilities/FunctionPointerStuff.h
@@ -23,10 +23,10 @@ void battle(Entity * e1, Entity * e2, int count, int * params);
 void enterLevel(Entity * e1, Entity * e2, int count, int * params);
 
 // functions to call on move
-void noMove(Entity * e, int c, int arr[2], int dt, int count, int * params);
+void noMove(Entity * e, int c, int arr[2], long dt, int count, int * params);
 
-void playerControl(Entity * e, int c, int arr[2], int dt, int count, int * params);
+void playerControl(Entity * e, int c, int arr[2], long dt, int count, int * params);
 
-void randomAI(Entity * e, int c, int arr[2], int dt, int count, int * params);
+void randomAI(Entity * e, int c, int arr[2], long dt, int count, int * params);
 
 #endif
